RAG/requirements.txt
# Core agentic frameworks
langchain
langgraph
langchain-core
langchain-google-genai

# Environment & validation
python-dotenv
pydantic
pyyaml

# Database
psycopg2-binary
sqlalchemy

# Embeddings / HuggingFace
sentence-transformers
huggingface_hub

# Notebook & visualization
ipython
graphviz
